# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Supernova API
  description: |
    REST API for the Supernova task management application.

    ## Overview
    Supernova is a task app where your lock screen wallpaper IS your task list,
    showing you the ONE thing to do right now.

    ## Authentication
    Most endpoints require JWT authentication via Bearer token in the Authorization header.
    - **Access Token**: 7-day expiry, used for API access
    - **Refresh Token**: 30-day expiry, used to obtain new access tokens
    - **Wallpaper Token**: 365-day expiry, limited scope for widget access

    ## Rate Limiting
    - Global: 100 req/min
    - Auth endpoints: 10 req/min
    - Wallpaper: 20 req/hour
    - Task creation: 30 req/min

  version: 1.0.0
  contact:
    name: Supernova Support
    url: https://github.com/supernova-app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.supernova.app
    description: Production server
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Create a new user account and receive authentication tokens
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: user@example.com
              password: SecurePass123!
              theme: cosmic
              timezone: America/New_York
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticate user and receive tokens
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: user@example.com
              password: SecurePass123!
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Exchange refresh token for new access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
                  description: Valid refresh token
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/wallpaper-token:
    post:
      tags:
        - Authentication
      summary: Generate wallpaper token
      description: Generate a new wallpaper token for widget access (invalidates old one)
      operationId: generateWallpaperToken
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  wallpaperToken:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================
  # USER ENDPOINTS
  # ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              message: Authentication required
              code: AUTHENTICATION_ERROR
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: Valid email address
        password:
          type: string
          minLength: 8
          description: Minimum 8 characters
        theme:
          type: string
          enum: [cosmic, ocean, fantasy]
          default: cosmic
        timezone:
          type: string
          description: IANA timezone
          default: UTC
        resolution:
          type: string
          pattern: '^\d+x\d+$'
          default: 1170x2532
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token (7-day expiry)
        refreshToken:
          type: string
          description: JWT refresh token (30-day expiry)
        wallpaperToken:
          type: string
          description: UUID token for widget access (1-year expiry)
        user:
          $ref: '#/components/schemas/UserBasic'

    # ----------------------------------------
    # User Schemas
    # ----------------------------------------
    UserBasic:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        theme:
          type: string
          enum: [cosmic, ocean, fantasy]
        timezone:
          type: string
        resolution:
          type: string
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - AUTHENTICATION_ERROR
                - AUTHORIZATION_ERROR
                - NOT_FOUND
                - CONFLICT
                - RATE_LIMIT
                - DATABASE_ERROR
                - CACHE_ERROR
                - INTERNAL_ERROR
            timestamp:
              type: string
              format: date-time
            path:
              type: string
            method:
              type: string
            details:
              type: object
    ValidationError:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
              msg:
                type: string
              path:
                type: string
              location:
                type: string